scripts/python/flaskapp/requirements.txt
Flask
pyopenssl
